test: 사다리 타기 관련된 기능(move) LadderGame -> Ladder 클래스로 옮김으로서 생긴 수정사항 수정
